pyproject.toml
[tool.poetry]
name = "temperature"
version = "0.1.0"
description = "Temperature sensor monitoring application"
authors = ["Moritz Enderle <contact@moritzenderle.com>"]
readme = "README.md"
packages = [{include = "temperature", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10"
redis = ">=7.0.1,<8.0.0"
fastapi = ">=0.121.1,<0.122.0"
sqlalchemy = ">=2.0.44,<3.0.0"
uvicorn = ">=0.30.0,<0.31.0"
pydantic = ">=2.5.0,<3.0.0"
tzdata = ">=2025.2,<2026.0"
pytz = ">=2025.2,<2026.0"
python-logging-loki = ">=0.3.1,<1.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
